change duration to integer

+++ database/migrations/2020_10_27_094155_create_course_semester_table.php
@@ -26,7 +26,7 @@ class CreateCourseSemesterTable extends Migration
                 ->references('id')->on('semesters')
                 ->cascadeOnDelete();
 
-            $table->decimal('duration', 2, 1);
+            $table->integer('duration');
 
             $table->timestampsTz();
